@@ -3,12 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "history": "^4.10.1",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.2.0",
-    "react-snow-effect": "^1.2.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "react-snowfall": "^1.2.1"
   },
   "scripts": {
     "start": "react-scripts start",
